Validation services consolidation
