+++ src/services.rs
@@ -3,7 +3,5 @@ pub mod command_validator;
 pub mod multi_package_installation_service;
 pub mod package_installation_service;
 pub mod package_list_service;
-pub mod package_validation_command;
-pub mod package_validation_service;
 pub mod package_validator;
-pub mod validation_service;
+pub mod validation_command;
