+++ tests/validation_integration_test.rs
@@ -5,7 +5,7 @@ use selfie::{
     adapters::progress::ProgressManager,
     domain::config::ConfigBuilder,
     ports::{command::MockCommandRunner, filesystem::MockFileSystem},
-    services::package_validation_command::{PackageValidationCommand, PackageValidationResult},
+    services::validation_command::{ValidationCommand, ValidationCommandResult},
 };
 use std::path::Path;
 
@@ -75,8 +75,7 @@ fn test_validation_integration() {
     let progress_manager = ProgressManager::new(false, false, false);
 
     // Create validation command
-    let command =
-        PackageValidationCommand::new(&fs, &runner, config.clone(), &progress_manager, false);
+    let command = ValidationCommand::new(&fs, &runner, config.clone(), &progress_manager, false);
 
     // Test validation on valid package
     let valid_cmd = selfie::domain::application::commands::PackageCommand::Validate {
@@ -86,7 +85,7 @@ fn test_validation_integration() {
 
     let result = command.execute(&valid_cmd);
     match result {
-        PackageValidationResult::Valid(output) => {
+        ValidationCommandResult::Valid(output) => {
             assert!(output.contains("valid-package"));
             assert!(output.contains("is valid"));
         }
@@ -101,7 +100,7 @@ fn test_validation_integration() {
 
     let result = command.execute(&invalid_cmd);
     match result {
-        PackageValidationResult::Invalid(output) => {
+        ValidationCommandResult::Invalid(output) => {
             eprintln!("{}", &output);
             assert!(output.contains("invalid-package"));
             assert!(output.contains("Validation failed"));
@@ -119,7 +118,7 @@ fn test_validation_integration() {
 
     let result = command.execute(&path_cmd);
     match result {
-        PackageValidationResult::Valid(_) => {
+        ValidationCommandResult::Valid(_) => {
             // Expected result
         }
         _ => panic!("Expected Valid result for path validation"),
